Address PR comments and fix unit test tolerance
