+++ src/Microsoft.ML.CpuMath/AlignedArray.cs
@@ -115,8 +115,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             Contracts.Assert(0 <= count && count <= _size);
             Contracts.Assert(dst != null);
             Contracts.Assert(0 <= index && index <= dst.Length - count);
-            for (int i = 0; i < count; i++)
-                dst[index + i] = Items[_base + i];
+            Items.AsSpan(_base, count).CopyTo(dst.Slice(index));
         }
 
         public void CopyTo(int start, Span<Float> dst, int index, int count)
@@ -125,8 +124,7 @@ namespace Microsoft.ML.Runtime.Internal.CpuMath
             Contracts.Assert(0 <= start && start <= _size - count);
             Contracts.Assert(dst != null);
             Contracts.Assert(0 <= index && index <= dst.Length - count);
-            for (int i = 0; i < count; i++)
-                dst[index + i] = Items[start + _base + i];
+            Items.AsSpan(start + _base).CopyTo(dst.Slice(index));
         }
 
         public void CopyFrom(ReadOnlySpan<Float> src)
